Move VectorIndexHnswCacheMemoryTracker to datashard
