+++ ydb/core/driver_lib/run/run.cpp
@@ -1580,9 +1580,6 @@ void TKikimrRunner::InitializeAppData(const TKikimrRunConfig& runConfig)
         AppData->DataShardConfig = runConfig.AppConfig.GetDataShardConfig();
     }
 
-    AppData->VectorIndexHnswCacheMemoryTracker->SetLimit(
-        AppData->DataShardConfig.GetVectorIndexHnswCacheMaxSize());
-
     if (runConfig.AppConfig.HasColumnShardConfig()) {
         AppData->ColumnShardConfig = runConfig.AppConfig.GetColumnShardConfig();
     }
