+++ ydb/core/tx/datashard/datashard.cpp
@@ -393,6 +393,8 @@ void TDataShard::OnActivateExecutor(const TActorContext& ctx) {
     LOG_INFO_S(ctx, NKikimrServices::TX_DATASHARD, "TDataShard::OnActivateExecutor: tablet " << TabletID() << " actor " << ctx.SelfID);
 
     InitControls();
+    VectorIndexHnswCacheMemoryTracker->SetLimit(
+        AppData(ctx)->DataShardConfig.GetVectorIndexHnswCacheMaxSize());
 
     // OnActivateExecutor might be called multiple times for a follower
     // but the counters should be initialized only once
